Tilemap Basic setup || Camera & Cage code changes
Made simple tilemap setup, met with few issues on the sizes of the tiles not matching with grid or one another. Made Camera trigger code to move the camera offset when needed and made Default offset value and current offset value for debugging. A quick change in Cage code to prevent multiple enemy spawns.